feat: ban

+++ service/src/main/java/com/prpo/chat/service/MembershipService.java
@@ -22,6 +22,7 @@ public class MembershipService {
 
     /**
      * Adds a user to the server (of type GROUP) as a member with a specific role
+     * Banned members cannot be added to the server.
      *
      * @param serverId the ID of the server to which the user will be added
      * @param userId   the ID of the user that will be added to the server
@@ -32,7 +33,14 @@ public class MembershipService {
             final String userId,
             final Membership.Role role) {
 
-        if (isMemberOfServer(userId, serverId)) {
+        final var membershipOptional = membershipRepository.findByServerIdAndUserId(serverId, userId);
+        if (membershipOptional.isPresent()) {
+            if(membershipOptional.get().getStatus() == Membership.Status.BANNED) {
+                throw new ResponseStatusException(
+                        HttpStatus.FORBIDDEN,
+                        String.format("User %s is banned from the server", userId)
+                );
+            }
             throw new ResponseStatusException(
                     HttpStatus.CONFLICT,
                     String.format("User %s is already member of the server %s", userId, serverId)
@@ -62,16 +70,20 @@ public class MembershipService {
 
     /**
      * Removes the targetUser from the server, if callerUser has a higher role.
-     * If targetUserId and callerUserId are the same, it means that that user wants to leave. the server.
+     * If targetUserId and callerUserId are the same, it means that that user wants to leave the server.
+     * Optionally bans user from the server.
      *
      * @param serverId ID of the server, can be of type GROUP or DM
      * @param callerUserId  ID of the user, that wants to kick the target user.
      * @param targetUserId ID of the user to be removed.
+     * @param isBanned if true, the status of the membership of the user is set to BANNED
      */
     public void removeMember(
             final String serverId,
             final String callerUserId,
-            final String targetUserId) {
+            final String targetUserId,
+            final boolean isBanned
+            ) {
 
         serverRepository.findById(serverId)
                 .orElseThrow(() -> new ResponseStatusException(
@@ -83,9 +95,15 @@ public class MembershipService {
                 .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
 
         /*
-         * Oseba lahko vedno odstrani samo sebe
+         * Oseba lahko vedno odstrani samo sebe (in v tem primeru ni banned)
          */
         if(callerUserId.equals(targetUserId)) {
+            if(isBanned) {
+                throw new ResponseStatusException(
+                        HttpStatus.FORBIDDEN,
+                        "User cannot ban themselves"
+                );
+            }
             // TODO: if the user is the OWNER, the OWNER role has to go to someone else
             membershipRepository.deleteByServerIdAndUserId(serverId, callerUserId);
             return;
@@ -100,12 +118,19 @@ public class MembershipService {
                     String.format("User %s does not have persmission to kick user %s.", callerUserId, targetUserId)
             );
         }
-        membershipRepository.deleteByServerIdAndUserId(serverId, targetUserId);
+
+        if(isBanned){
+            targetMembership.setStatus(Membership.Status.BANNED);
+            membershipRepository.save(targetMembership);
+        } else {
+            membershipRepository.deleteByServerIdAndUserId(serverId, targetUserId);
+        }
     }
 
     /**
      * Retrieves the list of servers that the given user is a member of,
      * optionally filtering by server type.
+     * Excludes servers the user is banned from.
      *
      * @param userId     the ID of the user whose servers are being fetched
      * @param serverType the type of servers to filter by (e.g. GROUP, DM);
@@ -114,10 +139,12 @@ public class MembershipService {
      */
     public List<Server> getServersForUser(final String userId, final Server.ServerType serverType) {
         final var serverIds = membershipRepository.findByUserId(userId).stream()
+                .filter(membership -> membership.getStatus() != Membership.Status.BANNED)
                 .map(Membership::getServerId)
                 .toList();
 
         var servers = serverRepository.findAllById(serverIds);
+
         if (serverType == null) {
             return servers;
         }
@@ -130,12 +157,14 @@ public class MembershipService {
 
     /**
      * Retrieves the list of users that are members of the given server
+     * Excluded users who are banned from the server.
      *
      * @param serverId the ID of the server
      * @return a list of userIds that are members of the server
      */
     public List<String> getUsersForServer(final String serverId) {
         return membershipRepository.findByServerId(serverId).stream()
+                .filter(membership -> membership.getStatus() != Membership.Status.BANNED)
                 .map(Membership::getUserId)
                 .toList();
     }
@@ -176,7 +205,4 @@ public class MembershipService {
         membershipRepository.save(membership);
     }
 
-    private boolean isMemberOfServer(final String userId, final String serverId) {
-        return membershipRepository.findByServerIdAndUserId(serverId, userId).isPresent();
-    }
 }
